feat: 实现自动注册驱动服务应用

+++ ruoyi-modules/ruoyi-manager/src/main/java/org/dromara/manager/driver/domain/DriverApplication.java
@@ -46,7 +46,7 @@ public class DriverApplication extends TenantEntity {
     /**
      * 端口号
      */
-    private Long port;
+    private Integer port;
 
 
 }
